test(prt): fix flowface/stopzone/extended tracking combo test
Fix some configuration to make sure PRT and MP7 are receiving identical flows. With this, endpoints now agree. There is still some divergence with extended tracking on, but it is small enough to chalk up to compiler toolchain/optimization flag differences.

Also fix the fix introduced in 2369 for termination in cells with zero flow under extended tracking, it should only apply if extended tracking is on. This deserves more consideration (can we avoid special casing terminations like this at the subcell level, while guaranteeing that tracking time is accurately reported?) in followup.