fix(prt-prp): fix release local z coord calculation
Fix conversion of local z release point coordinates to model z coordinates: take into account whether the cell is convertible or confined, and avoid incorrectly computing model z coordinates to be below the cell bottom. If convertible, define the cell's vertical extent from the bottom to the saturated top, clamping the latter to the bottom if head is below the cell bottom elevation; if confined, use the cell's full vertical extent from the bottom to the geometric top.